requirements.txt
tensorflow==2.7.0
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
streamlit==1.4.0
pandas==1.3.3
numpy==1.21.2
Pillow==8.4.0